defaultレイアウトのコンポーネント化
